btf: handle signed enums and fix default signedness
Enums now carry signedness encoded in the KindFlag, with the default being
an unsigned enum. This breaks our assumption that enums are signed by default.

Add Enum.Signed and switch EnumValue.Value to uint64. I used uint64 instead of
uint32 since we'll likely need 64 bits to support BTF_KIND_ENUM64 in the future.

Users of bpf2go that generated enums will see their generated types changing from
signed to unsigned types.

Updates #780

+++ btf/core_test.go
@@ -209,7 +209,7 @@ func TestCOREFindEnumValue(t *testing.T) {
 		name                    string
 		local, target           Type
 		acc                     coreAccessor
-		localValue, targetValue int32
+		localValue, targetValue uint64
 	}{
 		{"a to b", a, b, coreAccessor{0}, 23, 0},
 		{"b to a", b, a, coreAccessor{1}, 123, 42},
